validators.py file added, actions and main files modified to improve their code
